+++ buildVars.py
@@ -16,7 +16,7 @@ addon_info = {
 	# Translators: Long description to be shown for this add-on on add-on information from add-ons manager
 	"addon_description": _("""Provides enhanced terminal accessibility for Windows Terminal and PowerShell, enabling screen reader users to efficiently navigate and interact with command-line interfaces. Inspired by TDSR (Terminal Data Structure Reader) and incorporates functionality from both TDSR and Speakup. Advanced features inspired by community suggestions and discussions."""),
 	# version
-	"addon_version": "1.1.0",
+	"addon_version": "1.2.0",
 	# Author(s)
 	"addon_author": "Pratik Patel",
 	# URL for the add-on documentation support